// src/customers/interfaces/new-customer.interface.ts
export interface NewCustomerInitValue {
  custId: string;
  formId: string;
  branchId: string;
  displayBranchId: string;
  fullName: string;
  gender: string;
  placeOfBirth: string;
  dateOfBirth: string;
  emailAddress: string;
  phoneNumber: string;
  identityType: string;
  identityNumber: string;
  identityAddress: string;
  identityCity: string;
  identityZipCode: string;
  jobTitlePersonal: string;
  companyName: string;
  companyPhoneNumber: string;
  companyAddress: string;
  companyAddressCity: string;
  companyAddressZipCode: string;
  billingSalutation: string;
  billingName: string;
  billingPhone: string;
  billingEmail: string;
  billingJobTitle: string;
  technicalSalutation: string;
  technicalName: string;
  technicalPhone: string;
  technicalEmail: string;
  technicalJobTitle: string;
  installationAddress: string;
  installationAddressCity: string;
  installationAddressZipCode: string;
  packageCode: string;
  packageName: string;
  packagePrice: string;
  packageTop: string;
  promoId: string;
  ppn: string;
  salesId: string;
  managerSalesId: string;
  extendNote: string;
  approvalEmpId: string;
  npwpNumber: string;
  taxType: boolean;
  cetakDuluan: boolean;
}